+++ CHANGELOG.md
@@ -56,6 +56,7 @@
   - `unbase64`
 - Added support for specifying a schema string (including implicit struct syntax) when calling `DataFrame.create_dataframe`.
 - Added support for `DataFrameWriter.insert_into/insertInto`. This method also supports local testing mode.
+- Added support for `DataFrame.create_temp_view` to create a temporary view. It will fail if the view already exists.
 - Added support for multiple columns in the functions `map_cat` and `map_concat`.
 
 #### Experimental Features