+++ CHANGELOG.md
@@ -75,7 +75,8 @@
 - Added support for creating a temporary view via `DataFrame.create_or_replace_temp_view` from a DataFrame created by reading a file from a stage.
 - Added support for `value_contains_null` parameter to MapType.
 - Added `interactive` to telemetry that indicates whether the current environment is an interactive one.
-- Allow `session.file.get` in a Native App to read file paths starting with `/` from the current version 
+- Allow `session.file.get` in a Native App to read file paths starting with `/` from the current version
+- Added support for multiple aggregation functions after `DataFrame.pivot`.
 
 #### Bug Fixes