+++ CHANGELOG.md
@@ -54,6 +54,7 @@
   - `log10`
   - `percentile_approx`
   - `unbase64`
+- Added support for `seed` argument in `DataFrame.stat.sample_by`. Note that it only supports a `Table` object, and will be ignored for a `DataFrame` object.
 - Added support for specifying a schema string (including implicit struct syntax) when calling `DataFrame.create_dataframe`.
 - Added support for `DataFrameWriter.insert_into/insertInto`. This method also supports local testing mode.
 - Added support for `DataFrame.create_temp_view` to create a temporary view. It will fail if the view already exists.