+++ CHANGELOG.md
@@ -54,6 +54,8 @@
   - `log10`
   - `percentile_approx`
   - `unbase64`
+  - `normal`
+  - `randn`
 - Added support for `seed` argument in `DataFrame.stat.sample_by`. Note that it only supports a `Table` object, and will be ignored for a `DataFrame` object.
 - Added support for specifying a schema string (including implicit struct syntax) when calling `DataFrame.create_dataframe`.
 - Added support for `DataFrameWriter.insert_into/insertInto`. This method also supports local testing mode.