+++ CHANGELOG.md
@@ -74,6 +74,7 @@
 - Added support for `contains_null` parameter to ArrayType.
 - Added support for creating a temporary view via `DataFrame.create_or_replace_temp_view` from a DataFrame created by reading a file from a stage.
 - Added support for `value_contains_null` parameter to MapType.
+- Added support for using `Column` object in `Column.in_` and `functions.in_`. 
 - Added `interactive` to telemetry that indicates whether the current environment is an interactive one.
 - Allow `session.file.get` in a Native App to read file paths starting with `/` from the current version
 - Added support for multiple aggregation functions after `DataFrame.pivot`.